requirements.txt
pyjwt
flask_sqlalchemy
flasgger
python-dotenv
pymysql
cryptography